more refactoring/organising
